refactor(checklist): scope to react-ink, drop the core/react/ui layers
- the feature began as react-ink only; the core extraction plus the react primitives and the ui shadcn component existed solely to also ship it on web, which has no consumer yet
- move the ChecklistItemData/ChecklistData types and the useToolActivityChecklist hook into react-ink so the package owns them directly
- revert core, react, ui, the registry entry, and the web docs page to their pre-PR state
- changeset now covers react-ink only
