+++ .changeset/live-checklist.md
@@ -1,7 +1,5 @@
 ---
-"@assistant-ui/core": patch
-"@assistant-ui/react": patch
 "@assistant-ui/react-ink": patch
 ---
 
-feat: add shared checklist primitive. types and the useToolActivityChecklist hook live in core, unstyled web primitives in react, and react-ink reuses the core logic
+feat: add live checklist primitives (ChecklistPrimitive, LiveChecklist, DataChecklist) for tracking tool-call progress
