--- apps/docs/content/docs/primitives/checklist.mdx
@@ -1,209 +1,0 @@
----
-title: Checklist
-description: Live checklist that tracks tool-call progress across messages.
-platforms: ["react"]
----
-
-The Checklist displays a live-updating list of tool calls with their current status (pending, running, complete, error). It automatically derives items from the message's tool-call parts, showing users what the assistant is working on.
-
-`@assistant-ui/react` ships the unstyled building blocks: `ChecklistPrimitive` and the `useToolActivityChecklist` hook. The ready-to-use, styled `LiveChecklist` lives in `@assistant-ui/ui` (the shadcn component package), composed from those primitives.
-
-## Quick Start
-
-Drop the styled `LiveChecklist` inside your assistant message to show tool-call progress:
-
-```tsx
-import { LiveChecklist } from "@/components/assistant-ui/checklist";
-import { MessagePrimitive } from "@assistant-ui/react";
-
-function AssistantMessage() {
-  return (
-    <MessagePrimitive.Root>
-      <MessagePrimitive.Parts />
-      <LiveChecklist />
-    </MessagePrimitive.Root>
-  );
-}
-```
-
-`LiveChecklist` automatically reads the current message's tool-call parts and renders a checklist with status indicators. When no tool calls are present, it renders nothing.
-
-<Callout type="info">
-Runtime setup: primitives require runtime context. Wrap your UI in `AssistantRuntimeProvider` with a runtime (for example `useLocalRuntime(...)`). See [Pick a Runtime](/docs/runtimes/pick-a-runtime).
-</Callout>
-
-## Styled Component
-
-The `@assistant-ui/ui` `LiveChecklist` renders status indicators (□/■/✕), a shimmer animation on running items, and an optional progress counter.
-
-### Auto-Derived Mode
-
-When used without an `items` prop, `LiveChecklist` reads tool-call parts from the current message context and maps them to checklist items:
-
-- Tool calls without a result are **running** (first) or **pending** (rest)
-- Tool calls with a result are **complete**
-- Tool calls with errors are marked as **error**
-
-Only the first incomplete tool call is shown as "running"; the rest are "pending". This creates the sequential progress effect.
-
-### Standalone Mode
-
-Pass an `items` prop to render a checklist with manually controlled data:
-
-```tsx
-import { LiveChecklist } from "@/components/assistant-ui/checklist";
-import type { ChecklistItemData } from "@assistant-ui/react";
-
-const items: ChecklistItemData[] = [
-  { id: "1", text: "Search the web", status: "complete" },
-  { id: "2", text: "Analyze results", status: "running" },
-  { id: "3", text: "Summarize findings", status: "pending" },
-];
-
-<LiveChecklist items={items} showProgress />
-```
-
-### Formatting Tool Names
-
-By default, tool names are displayed as-is. Use `formatToolName` to transform them:
-
-```tsx
-<LiveChecklist formatToolName={(name) => name.replace(/_/g, " ")} />
-```
-
-| Prop | Type | Default | Description |
-|------|------|---------|-------------|
-| `items` | `ChecklistItemData[]` | — | Manual items (omit for auto mode) |
-| `showProgress` | `boolean` | `true` | Show "X/Y complete" counter |
-| `formatToolName` | `(name: string) => string` | — | Transform tool names for display |
-| `title` | `ReactNode` | — | Optional title above the list |
-| `maxDepth` | `number` | `2` | Maximum nesting depth |
-| `variant` | `"outline" \| "ghost" \| "muted"` | `"outline"` | Container style |
-
-### DataChecklist
-
-Renders checklist data sent via `makeAssistantDataUI`. Register `ChecklistDataUI` so the runtime renders incoming `checklist` data parts:
-
-```tsx
-import { DataChecklist, ChecklistDataUI } from "@/components/assistant-ui/checklist";
-```
-
-The progress counter is hidden by default; set `showProgress: true` in the `ChecklistData` payload to display it.
-
-## Primitives
-
-For full control over markup and styling, compose the unstyled primitives from `@assistant-ui/react` directly.
-
-### ChecklistPrimitive.Root
-
-Container for the checklist. Renders a `<div>`.
-
-```tsx
-import { ChecklistPrimitive } from "@assistant-ui/react";
-
-<ChecklistPrimitive.Root className="flex flex-col gap-1">
-  ...
-</ChecklistPrimitive.Root>
-```
-
-### ChecklistPrimitive.Item
-
-Renders a single checklist item as a `<div>` with a `data-status` attribute. By default it renders the item's text (plus a `data-detail` span when `detail` is set); pass `children` to fully control the row's content.
-
-```tsx
-<ChecklistPrimitive.Item
-  item={{ id: "1", text: "Search", status: "running" }}
-/>
-```
-
-| Prop | Type | Default | Description |
-|------|------|---------|-------------|
-| `item` | `ChecklistItemData` | required | The checklist item data |
-| `children` | `ReactNode` | — | Custom row content (replaces the default text markup) |
-
-The primitive renders one row. Walk the tree yourself to render nested `children` (the styled `LiveChecklist` does this up to `maxDepth`):
-
-```tsx
-<ChecklistPrimitive.Item item={item}>
-  <StatusIcon status={item.status} />
-  <span>{item.text}</span>
-</ChecklistPrimitive.Item>
-```
-
-### ChecklistPrimitive.Progress
-
-Renders progress as "X/Y complete". Renders a `<div>` with `data-done` and `data-total` attributes.
-
-```tsx
-<ChecklistPrimitive.Progress items={items} />
-```
-
-## Hook
-
-### useToolActivityChecklist
-
-Returns `ChecklistItemData[]` derived from the current message's tool-call parts. This is the logic the styled `LiveChecklist` uses for its auto-derived mode:
-
-```tsx
-import { useToolActivityChecklist } from "@assistant-ui/react";
-
-function MyChecklist() {
-  const items = useToolActivityChecklist({
-    formatToolName: (name) => name.replace(/_/g, " "),
-  });
-
-  return items.map((item) => (
-    <div key={item.id} data-status={item.status}>
-      {item.text}
-    </div>
-  ));
-}
-```
-
-## Types
-
-```tsx
-type ChecklistItemStatus = "pending" | "running" | "complete" | "error";
-
-type ChecklistItemData = {
-  id: string;
-  text: string;
-  status: ChecklistItemStatus;
-  detail?: string;
-  children?: ChecklistItemData[];
-};
-
-type ChecklistData = {
-  items: ChecklistItemData[];
-  title?: string;
-  showProgress?: boolean;
-};
-```
-
-## Patterns
-
-### Inside AssistantMessage
-
-```tsx
-function AssistantMessage() {
-  return (
-    <MessagePrimitive.Root>
-      <div className="flex flex-col gap-2">
-        <MessagePrimitive.Parts />
-        <LiveChecklist
-          showProgress
-          formatToolName={(name) => name.replace(/_/g, " ")}
-        />
-      </div>
-    </MessagePrimitive.Root>
-  );
-}
-```
-
-## API Reference
-
-For prop details, see the [ChecklistPrimitive source](https://github.com/assistant-ui/assistant-ui/tree/main/packages/react/src/primitives/checklist).
-
-Related:
-- [MessagePrimitive](/docs/primitives/message)
-- [ChainOfThoughtPrimitive](/docs/primitives/chain-of-thought)
