+++ packages/core/src/index.ts
@@ -61,12 +61,6 @@ export type { Unsubscribe } from "./types/unsubscribe";
 export type { QuoteInfo } from "./types/quote";
 
 export type {
-  ChecklistItemData,
-  ChecklistItemStatus,
-  ChecklistData,
-} from "./types/checklist";
-
-export type {
   Unstable_DirectiveSegment,
   Unstable_DirectiveFormatter,
 } from "./types/directive";
