+++ packages/core/src/react/index.ts
@@ -254,10 +254,6 @@ export { useThreadListLoadMore } from "./primitive-hooks/useThreadListLoadMore";
 export { useEditComposerCancel } from "./primitive-hooks/useEditComposerCancel";
 export { useEditComposerSend } from "./primitive-hooks/useEditComposerSend";
 export { useMessageError } from "./primitive-hooks/useMessageError";
-export {
-  useToolActivityChecklist,
-  type UseToolActivityChecklistOptions,
-} from "./primitive-hooks/useToolActivityChecklist";
 
 // Shared AssistantRuntimeProvider
 export { AssistantRuntimeProvider } from "./AssistantRuntimeProvider";
