--- packages/react/src/primitives/checklist.ts
@@ -1,5 +1,0 @@
-export { ChecklistPrimitiveRoot as Root } from "./checklist/ChecklistRoot";
-
-export { ChecklistPrimitiveItem as Item } from "./checklist/ChecklistItem";
-
-export { ChecklistPrimitiveProgress as Progress } from "./checklist/ChecklistProgress";
